backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
chromadb==0.5.11
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
httpx==0.27.2
